PlatformIO/vest/pyproject.toml
[project]
name = "test-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleak>=2.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pyserial>=3.5",
    "pyvista>=0.47.1",
    "scipy>=1.17.1",
]
